Extract events RBAC access and cache logic into dedicated modules
Split `events.ts` into `eventsAccess.ts` for RBAC evaluation and
`eventsCache.ts` for access cache management, and add unit tests for
both new modules.
